cms-dservice/requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
certifi==2018.10.15
chardet==3.0.4
diff-match-patch==20121119
dill==0.3.1.1
Django==2.1.3
django-admin-rangefilter==0.3.8
django-bootstrap4==0.0.7
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==2.4.0
django-filter==2.0.0
django-import-export==1.1.0
django-jet==1.0.7
django-oauth-toolkit==1.2.0
django-prometheus==1.0.15
django-simple-history==2.7.0
django-timezone-field==3.0
djangorestframework==3.9.0
dynamic-rest==1.8.1
et-xmlfile==1.0.1
idna==2.7
inflection==0.3.1
jdcal==1.4
kombu==4.6.1
mysqlclient==1.3.13
nltk==3.4.5
numpy==1.18.1
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.8
pandas==1.0.1
Pillow==5.3.0
prometheus-client==0.7.1
pycrypto==2.6.1
pythainlp==2.1.3
python-crontab==2.3.5
python-dateutil==2.7.5
pytz==2018.5
PyYAML==3.13
redis==3.2.1
regex==2018.8.29
reportlab==3.5.9
requests==2.22.0
six==1.11.0
tablib==0.12.1
tinydb==3.15.2
tqdm==4.41.1
unicodecsv==0.14.1
urllib3==1.23
vine==1.3.0
xlrd==1.1.0
xlwt==1.3.0
